Add specific code binding information to Data Elements
When a value in the "Data Elements" worksheet is constrained to a specific code or value set, indicate this in the "Terminology Binding" column. In the case of a specific code binding, provide information about that code in the "Code System", "Code", and "Code Description" columns. In the case of a value set binding, provide information about that value set in the "Value Set" and "Value Set Binding" columns.