Terminology Binding empty if no code or value set, reorder columns in Value Set Details
When a data element displayed in the Data Elements tab has no associated
fixed code or value set, the output for the Terminology Binding should
be empty.

In Value Set Details tab, the Logical Definition column has been inserted as
the third column. Code is now the fourth column, and Code Description is now
the fifth column.